Type in VarNode changed to RestoredMemory
